+++ science/__init__.py
@@ -3,6 +3,6 @@
 
 from packaging.version import Version
 
-__version__ = "0.6.1"
+__version__ = "0.7.0"
 
 VERSION = Version(__version__)